requirements.txt


scikit-learn==1.5.1
Pillow==10.4.0
pytube
youtube-dl
streamlit-tags
pyresparser
pdfminer3
pymysql
plotly
streamlit==1.28.1
streamlit-option-menu==0.3.6
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
pandas==2.0.3
plotly==5.17.0
openpyxl==3.1.2
requests==2.28.0